fix: moved public route resolver into ui package
