+++ packages/plugins/public-layout-ui/src/lib/public-api.ts
@@ -1,5 +1,8 @@
 /**
  * Public API Surface of @gauzy/plugin-public-layout-ui
  */
+export * from './components/public-layout/public-layout.component';
 export * from './public-layout.module';
 export * from './public-layout.routes';
+export * from './resolvers/public-employee.resolver';
+export * from './resolvers/public-organization.resolver';
